+++ tests/vector_tests/vector_push_pop_back_tests.cpp
@@ -80,7 +80,7 @@ TEMPLATE_TEST_CASE_SIG("Push back many times", "[vector] [push]",
 		REQUIRE(vector_number.size() == ft_vector_number.size());
 		REQUIRE(vector_number.front() == ft_vector_number.front());
 		REQUIRE(vector_number.back() == ft_vector_number.back());
-		for (int i = 0; i < vector_number.size(); ++i) {
+		for (size_t i = 0; i < vector_number.size(); ++i) {
 			REQUIRE(ft_vector_number[i] == vector_number[i]);
 		}
 	}
@@ -144,7 +144,7 @@ TEMPLATE_TEST_CASE_SIG("Pop back many times", "[vector] [pop]",
 		REQUIRE(vector_number.size() == ft_vector_number.size());
 		REQUIRE(vector_number.front() == ft_vector_number.front());
 		REQUIRE(vector_number.back() == ft_vector_number.back());
-		for (int i = 0; i < vector_number.size(); ++i) {
+		for (size_t i = 0; i < vector_number.size(); ++i) {
 			REQUIRE(ft_vector_number[i] == vector_number[i]);
 		}
 	}
@@ -163,7 +163,7 @@ TEMPLATE_TEST_CASE_SIG("Pop back many times", "[vector] [pop]",
 		REQUIRE(vector_string.size() == ft_vector_string.size());
 		REQUIRE(vector_string.front() == ft_vector_string.front());
 		REQUIRE(vector_string.back() == ft_vector_string.back());
-		for (int i = 0; i < vector_string.size(); ++i) {
+		for (size_t i = 0; i < vector_string.size(); ++i) {
 			REQUIRE(ft_vector_string[i] == vector_string[i]);
 		}
 	}
